# ♻️ Refactor: reduce DefaultCtx from 768 bytes to 736 bytes