# Description

In the default context, `pathBuffer` and `detectionPathBuffer` can be used as `path` and `detectionPath` directly. We can remove two `string` fields and save `32 bytes` for each context.

```
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v3
cpu: AMD EPYC 9J14 96-Core Processor                
                                             │   old.txt   │              new.txt               │
                                             │   sec/op    │   sec/op     vs base               │
_Ctx_AcquireReleaseFlow/withoutRequestCtx-16   41.30n ± 1%   39.10n ± 0%  -5.33% (p=0.000 n=25)
_Ctx_AcquireReleaseFlow/withRequestCtx-16      530.8n ± 1%   518.7n ± 1%  -2.28% (p=0.000 n=25)
geomean                                        148.1n        142.4n       -3.82%

                                             │   old.txt    │                new.txt                │
                                             │     B/op     │     B/op      vs base                 │
_Ctx_AcquireReleaseFlow/withoutRequestCtx-16     3.000 ± 0%     3.000 ± 0%       ~ (p=1.000 n=25) ¹
_Ctx_AcquireReleaseFlow/withRequestCtx-16      1.532Ki ± 0%   1.532Ki ± 0%       ~ (p=1.000 n=25) ¹
geomean                                          68.61          68.61       +0.00%
¹ all samples are equal

                                             │  old.txt   │               new.txt               │
                                             │ allocs/op  │ allocs/op   vs base                 │
_Ctx_AcquireReleaseFlow/withoutRequestCtx-16   1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=25) ¹
_Ctx_AcquireReleaseFlow/withRequestCtx-16      5.000 ± 0%   5.000 ± 0%       ~ (p=1.000 n=25) ¹
geomean                                        2.236        2.236       +0.00%
¹ all samples are equal

```

## Type of change

- [x] Enhancement (improvement to existing features and functionality)
- [x] Code consistency (non-breaking change which improves code reliability and robustness)

## Checklist

Before you submit your pull request, please make sure you meet these requirements:

- [ ] Followed the inspiration of the Express.js framework for new functionalities, making them similar in usage.
- [x] Conducted a self-review of the code and provided comments for complex or critical parts.
- [ ] Updated the documentation in the `/docs/` directory for [Fiber's documentation](https://docs.gofiber.io/).
- [ ] Added or updated unit tests to validate the effectiveness of the changes or new features.
- [x] Ensured that new and existing unit tests pass locally with the changes.
- [ ] Verified that any new dependencies are essential and have been agreed upon by the maintainers/community.
- [x] Aimed for optimal performance with minimal allocations in the new code.
- [ ] Provided benchmarks for the new code to analyze and improve upon.