@ReneWerner87 2025-03-19
I am still unsure, at that time the additional buffers were necessary to reduce the allocations and preserve the original
can we find the case from back then and check it again?