@ReneWerner87 2025-03-20
> > I am still unsure, at that time the additional buffers were necessary to reduce the allocations and preserve the original can we find the case from back then and check it again?
> 
> I reduce the two string fields in the `defautCtx` and keep the original two buffers. I think the `c.detectionPath` and `c.path` can be saved by `[]byte` in the sturct. They can be lazy converted to the string until the user calls c.Path() or c.detectPath().

yes already understood, just want to make sure that we don't have any performance losses(more allocs) or that the strings are not changed by the referencing

I will test this by hand
then merge, have to make sure again, because there was a case back then