@ReneWerner87 2025-03-23
really good work, I checked everything again and we also have tests for the cases where I suspected that something could still go wrong

https://github.com/ksw2000/fiber/blob/395c8fafa96719cfcce808876c4b11914ca34efc/app_test.go#L500

thanks for your work and patience